Fixed the diagnostics mode for >12v bus voltages.  Had to change the BusVoltage variable from a WORD to DWORD to handle possible averaging filter overflows.   Cleaned up alot of the other spport files.  Tried to change variable declarations to DataType.h macros

+++ SRC/DEVELOPMENT_RELEASE/MASTER/COMMON/INCLUDE/ADCRoutines.h
@@ -1,8 +1,11 @@
+#include "DataTypes.h"
+
+
 #ifndef _ADCROUTINES_H
 #define _ADCROUTINES_H
 
 void ADCInit(void);
-unsigned int ADCRead(void);
+WORD ADCRead(void);
 
 #endif
